pyproject.toml
[project]
name = "ai-llm-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "openai>=2.24.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
